auto format all files

+++ token/contracts/TokenSale.sol
@@ -9,24 +9,23 @@ contract TokenSale {
 
     address private _admin;
     Token public tokenContract;
-    uint public tokenPrice;
-    uint public tokenSold;
+    uint256 public tokenPrice;
+    uint256 public tokenSold;
 
-    event Sell(address _buyer, uint _amount);
+    event Sell(address _buyer, uint256 _amount);
 
-    constructor(Token _tokenContract, uint _tokenPrice) public {
+    constructor(Token _tokenContract, uint256 _tokenPrice) public {
         _admin = msg.sender;
         tokenContract = _tokenContract;
         tokenPrice = _tokenPrice;
     }
 
-    function buyTokens(uint _numTokens) public payable {
+    function buyTokens(uint256 _numTokens) public payable {
         require(msg.value == _numTokens.mul(tokenPrice));
         require(tokenContract.balanceOf(address(this)) >= _numTokens);
         require(tokenContract.transfer(msg.sender, _numTokens));
         tokenSold += _numTokens;
-        
+
         emit Sell(msg.sender, _numTokens);
     }
 }
-
